+++ app/src/main/java/org/sopt/do_sopt_compose/ui/components/Birthday.kt
@@ -17,30 +17,29 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import androidx.core.graphics.toColorInt
 import org.sopt.do_sopt_compose.R
 
 @Composable
-fun MelonMusic(
-    music: String,
+fun Birthday(
 ) {
-
     Card(
         modifier = Modifier
             .wrapContentSize()
             .padding(0.dp, 12.dp),
         shape = RoundedCornerShape(corner = CornerSize(16.dp)),
-        border = BorderStroke(1.dp, Color.Green)
+        border = BorderStroke(1.dp, Color.Red)
     ) {
         Row(modifier = Modifier.padding(8.dp)) {
             Image(
-                painter = painterResource(id = R.drawable.ic_android_24dp),
+                painter = painterResource(id = R.drawable.baseline_cake_24),
                 contentDescription = null,
                 alignment = Alignment.Center
             )
             Text(
-                text = music,
+                text = "생일인 친구",
                 style = MaterialTheme.typography.bodySmall,
-                color = Color.Green,
+                color = Color.Red,
                 modifier = Modifier.padding(8.dp)
             )
         }
