+++ app/src/main/java/org/sopt/do_sopt_compose/ui/components/profile/FriendsProfileBirthday.kt
@@ -1,10 +1,11 @@
-package org.sopt.do_sopt_compose.ui.components
+package org.sopt.do_sopt_compose.ui.components.profile
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.CircleShape
@@ -17,7 +18,9 @@ import androidx.compose.ui.graphics.ImageBitmap
 import androidx.compose.ui.res.imageResource
 import androidx.compose.ui.text.font.FontStyle
 import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
+import org.sopt.do_sopt_compose.ui.components.Birthday
 
 
 @Composable
@@ -28,7 +31,7 @@ fun FriendsProfileBirthday(
 ) {
     Box(
         modifier = Modifier
-            .fillMaxSize()
+            .fillMaxWidth()
             .padding(5.dp)
     ) {
         Row {
@@ -42,11 +45,10 @@ fun FriendsProfileBirthday(
 
             Column {
                 Text(
-                    text = "오늘 생일인$nickname",
+                    text = nickname,
                     textAlign = TextAlign.Center,
                     fontStyle = FontStyle.Normal,
                     modifier = Modifier.padding(start = 8.dp),
-                    color = Color.Red
                 )
 
                 Text(
@@ -54,8 +56,11 @@ fun FriendsProfileBirthday(
                     textAlign = TextAlign.Center,
                     fontStyle = FontStyle.Italic,
                     modifier = Modifier.padding(start = 8.dp),
-                    maxLines = 1
+                    overflow = TextOverflow.Ellipsis,
+                    maxLines = 1,
                 )
+
+                Birthday()
             }
         }
     }
