+++ app/src/main/java/org/sopt/do_sopt_compose/ui/components/profile/FriendsProfileNormal.kt
@@ -1,9 +1,10 @@
-package org.sopt.do_sopt_compose.ui.components
+package org.sopt.do_sopt_compose.ui.components.profile
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.CircleShape
@@ -15,35 +16,46 @@ import androidx.compose.ui.graphics.ImageBitmap
 import androidx.compose.ui.res.imageResource
 import androidx.compose.ui.text.font.FontStyle
 import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 
-
 @Composable
-fun FriendsProfileMe(
+fun FriendsProfileNormal(
     profileImageUrl: Int,
     nickname: String,
+    statusMessage: String?,
 ) {
     Box(
         modifier = Modifier
-            .fillMaxSize()
-            .padding(5.dp)
+            .fillMaxWidth()
+            .padding(10.dp)
     ) {
         Row {
             Image(
                 bitmap = ImageBitmap.imageResource(profileImageUrl),
                 contentDescription = null,
                 modifier = Modifier
-                    .size(80.dp)
+                    .size(50.dp)
                     .clip(CircleShape)
             )
 
-            Text(
-                text = nickname,
-                textAlign = TextAlign.Center,
-                fontStyle = FontStyle.Normal,
-                modifier = Modifier
-                    .padding(start = 8.dp)
-            )
+            Column {
+                Text(
+                    text = nickname,
+                    textAlign = TextAlign.Center,
+                    fontStyle = FontStyle.Normal,
+                    modifier = Modifier.padding(start = 8.dp)
+                )
+
+                Text(
+                    text = statusMessage ?: "",
+                    textAlign = TextAlign.Center,
+                    fontStyle = FontStyle.Italic,
+                    modifier = Modifier.padding(start = 8.dp),
+                    overflow = TextOverflow.Ellipsis,
+                    maxLines = 1,
+                )
+            }
         }
     }
 }
