+++ app/src/main/res/drawable/ic_home_24.xml
@@ -1,0 +1,5 @@
+<vector android:height="24dp" android:tint="#B0B0B0"
+    android:viewportHeight="24" android:viewportWidth="24"
+    android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="@android:color/white" android:pathData="M10,20v-6h4v6h5v-8h3L12,3 2,12h3v8z"/>
+</vector>
